pyproject.toml
[tool.poetry]
name = "school-bot"
version = "0.1.0"
description = "telegram agrigator-bot for sending links to zoom lesson meetings"
authors = ["farlo0nn <sasha.b0707@gmail.com>"]
readme = "README.md"
packages = [{include = "school_bot"}]

[tool.poetry.dependencies]
python = "3.10.10"
asn1crypto = "0.24.0"
black = "23.1.0"
certifi = "2022.9.24"
cffi = "1.15.1"
charset-normalizer = "2.1.1"
click = "8.1.3"
colorama = "0.4.6"
cryptography = "2.1.4"
enum34 = "1.1.6"
flask = "2.2.2"
flask-wtf = "1.1.1"
greenlet = "2.0.1"
idna = "2.6"
ipaddress = "1.0.17"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
keyring = "10.6.0"
keyrings-alt = "3.0"
loguru = "0.6.0"
markupsafe = "2.1.2"
mypy-extensions = "1.0.0"
packaging = "23.0"
pathspec = "0.11.0"
platformdirs = "3.0.0"
pycparser = "2.21"
pytelegrambotapi = "4.7.1"
python-decouple = "3.7"
python-dotenv = "0.21.0"
pywin32-ctypes = "0.2.0"
pyxdg = "0.25"
requests = "2.28.1"
schedule = "1.1.0"
secretstorage = "2.3.1"
six = "1.11.0"
sqlalchemy = "1.4.46"
tomli = "2.0.1"
urllib3 = "1.26.12"
werkzeug = "2.2.2"
win32-setctime = "1.1.0"
wtforms = "3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
